Fix bug

+++ commands/admin/adminHp.js
@@ -8,7 +8,7 @@ const AdminManager  = require("../../classes/AdminManager");
 class AdminHpCommand extends Command {
    constructor() {
       super('adminHp', {
-         aliases: ['setBoss'],
+         aliases: ['setbosshp'],
          cooldown: 3000,
          channel: 'guild',
          userPermissions: Permissions.FLAGS.ADMINISTRATOR,
@@ -17,6 +17,14 @@ class AdminHpCommand extends Command {
                id: 'compensate',
                match: 'flag',
                flag: ['c','-c','com','-com']
+            },
+			{
+               id: "boss",
+               type: "integer",
+               prompt: {
+                  start: strings.prompt.boss,
+                  retry: strings.prompt.not_a_number,
+               },
             },
             {
                id: "hp",
@@ -31,28 +39,32 @@ class AdminHpCommand extends Command {
    };
 
    async exec(message, args) {
-      const guildID = message.guild.id
-      const clientID = message.author.id
-      const db = this.client.db
-      const dm = new AdminManager(db,guildID,clientID)
+      	const guildID = message.guild.id
+      	const clientID = message.author.id
+      	const db = this.client.db
+      	const dm = new AdminManager(db,guildID)
       
-      let loadingMsg = await message.channel.send(strings.common.waiting);
+      	let loadingMsg = await message.channel.send(strings.common.waiting);
 
-      const hp = args.hp
+      	const hp = args.hp
+ 	  	const boss = args.boss
 
-      const detail = await dm.resetBossHp(hp)
+		if (boss < 1 || boss > 5) {
+    		loadingMsg.edit(strings.common.boss_out_of_range);
+      		return
+    	};
+
+      	const detail = await dm.resetBossHp(boss, hp)
       
-      const embed = new MessageEmbed();
-      embed.setColor("#90ffff");
-      embed.setTitle(admin.reset.hp.replace('[hp]', hp));
-      loadingMsg.delete()
-      await message.channel.send(embed);
-      this.client.emit("reportUpdate", message.guild);
+      	const embed = new MessageEmbed();
+      	embed.setColor("#90ffff");
+      	embed.setTitle(admin.reset.hp.replace('[hp]', hp).replace('[boss]', boss));
+      	loadingMsg.delete()
+      	await message.channel.send(embed);
+      	this.client.emit("reportUpdate", message.guild);
 
-      return
+      	return
     }
-    
- 
 }
 module.exports = AdminHpCommand;
